hide create-private type from public methods
